pyproject.toml
[tool.poetry]
name = "work-schedule"
version = "0.1.0"
description = ""
authors = ["lordtachen <matos.aes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
pyyaml = "^6.0.1"
pip = "^23.3.2"
install = "^1.3.5"
pydantic = {extras = ["email"], version = "^2.5.2"}
httpx = "^0.25.2"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
